requirements.txt
# Core frameworks
# NOTE: Requires Python 3.10 (lava-nc constraint)
lava-nc==0.10.0
torch>=2.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
scipy>=1.10.0,<2.0.0

# Configuration and utilities
pyyaml>=6.0
psutil>=5.9.0

# Testing
pytest>=7.0.0
